requirements.txt

# Additional requirements go here

djangorestframework==3.7.7

django-filter==1.1.0
django-genomix==0.6.0
django-model-utils==3.1.1

graphene-django==2.0.0
pronto==0.10.2
